feature: all scripts are now Makefile relative
Each of the previoulsy written scripts now use shell variables. Making
them not executable alone, but through a newly created Makefile.
The makefile defines rules for deploying a Kong-Docker environement
but also for packaging the plugin and installing it in our Kong env.
It now requires only one 'make all' to ensure a running Kong env, and
a deployed plugin.